requirements.txt
annotated-types==0.6.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.2
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.6.4
pydantic-core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
tqdm==4.66.2
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.2
